+++ Classes/00-Base/WSUSPackageDefinition.ps1
@@ -1,0 +1,16 @@
+class WSUSPackageDefinition : Package {
+    [string]$UpdateId
+    [int]$RevisionNumber
+    [string[]]$KBArticleIds
+    [string]$ApprovalStatus
+    [bool]$Applicable
+    [bool]$Downloaded
+    [bool]$Mandatory
+
+    WSUSPackageDefinition() {
+        $this.Provider = 'WSUS'
+        $this.InstallerType = 'WindowsUpdate'
+        $this.RequiresElevation = $true
+        $this.KBArticleIds = @()
+    }
+}